Revert "try fix round"
This reverts commit adf873a752220b2c51da3f02f8a95c513fc19956.

+++ src/main/java/tkachgeek/tkachutils/numbers/NumbersUtils.java
@@ -29,7 +29,7 @@ public class NumbersUtils {
     * @return округляет value до нужного количества цифр (digits) после запятой
     */
    public static double round(double value, int digits) {
-      int scale = (int) Math.pow(10, digits);
+      double scale = Math.pow(10, digits);
       return Math.ceil(value * scale) / scale;
    }
